backend/requirements.txt
fastapi>=0.110
python-multipart>=0.0.9
uvicorn[standard]>=0.27
python-dotenv>=1.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
email-validator>=2.0
sqlalchemy[asyncio]>=2.0
aiomysql>=0.2
aiosqlite>=0.20
greenlet>=3.0
openai>=1.0
numpy>=1.24
sentence-transformers>=2.6
torch>=2.0
transformers>=4.40
